requirements.txt
pyYAML==6.0.3
open3d==0.19.0
numba==0.63.1
scipy==1.17.0
ipywidgets==7.7.5
jupyterlab-widgets==1.1.1
opencv-contrib-python==4.13.0.92
scikit-learn==1.8.0
scikit-image==0.26.0
matplotlib==3.10.8
tqdm==4.67.3
laspy==2.7.0
ezdxf==1.4.3
dxfwrite==1.2.2
setuptools==82.0.0
tkinter-embed==3.12.0
nbclassics==1.3.3

#Old PBS requirements
cgal==6.0.1.post202410241521
PyQt5==5.15.11
descartes==1.1.0
vispy==0.16.1
shapely==2.1.2